Returned original setup to integer

+++ backend/src/main/java/ba/atlant/auctionapp/service/PersonService.java
@@ -91,8 +91,8 @@ public class PersonService {
     }
 
     public ResponseEntity<PersonProjection> getCurrentUser(String authHeader) {
-        Long userId = jwtUtils.getUserIdFromJwtToken(authHeader.substring(7));
-        personRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("No user found with provided ID."));
+        Integer userId = jwtUtils.getUserIdFromJwtToken(authHeader.substring(7));
+        personRepository.findById(Long.valueOf(userId)).orElseThrow(() -> new ResourceNotFoundException("No user found with provided ID."));
         PersonProjection personProjection = personRepository.getPersonInformation(Long.valueOf(userId.toString()));
         return ResponseEntity.ok(personProjection);
     }
@@ -130,7 +130,7 @@ public class PersonService {
     @Transactional
     public ResponseEntity<PersonDTO> modifyCurrentUser(String authHeader, PersonDTO personDTO) {
         try {
-            Long userId = jwtUtils.getUserIdFromJwtToken(authHeader.substring(7));
+            Long userId = Long.valueOf(jwtUtils.getUserIdFromJwtToken(authHeader.substring(7)));
             Person person = personRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("No user found with provided ID."));
             person.setFirstName(personDTO.getFirstName());
             person.setLastName(personDTO.getLastName());
@@ -153,7 +153,7 @@ public class PersonService {
 
     @Transactional
     public ResponseEntity<Person> addPictureToUser(String authHeader, MultipartFile file) throws IOException {
-        Long userId = jwtUtils.getUserIdFromJwtToken(authHeader.substring(7));
+        Long userId = Long.valueOf(jwtUtils.getUserIdFromJwtToken(authHeader.substring(7)));
         Person person = personRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("No user found with provided ID."));
         s3Service.deleteObject(person.getPictureUrl());
         s3Service.uploadFile(file.getOriginalFilename(), file);
@@ -163,7 +163,7 @@ public class PersonService {
     }
 
     public ResponseEntity<Map<String, String>> getUserPicture(String authHeader) {
-        Long userId = jwtUtils.getUserIdFromJwtToken(authHeader.substring(7));
+        Long userId = Long.valueOf(jwtUtils.getUserIdFromJwtToken(authHeader.substring(7)));
         Person person = personRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("No user found with provided ID."));
         Map<String, String> personPictureUrl = new HashMap<>();
         personPictureUrl.put("url", person.getPictureUrl());
